Modularized extract, ignoring checkpoints
the extraction of files is now modular (for cleaner code). Now it will try to
extract from new_files if it exists in order to extract from new searches,
otherwise it will extract from all files.

parse_backup_info now discards checkpoints because they are incomplete. This
is particularly a problem because whether or not a backup is a checkpoint is
not apparent to the user.

Signed-off-by: Rick Henry <fredericmhenry@gmail.com>